TriNetra/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.0
numpy<2
scikit-learn==1.3.0
python-dateutil==2.8.2
Faker==19.6.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
networkx==3.1
shap==0.42.1
python-louvain
supabase>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0